Fix rare test flakiness
The `filter` method is a boolean check that removes items that are not
truthy. The previous behavior would evaluate the example as a boolean
after running `strip()`. Thus removing empty strings or just spaces.
However, it didnt actually change the values, meaning examples with
preceeding or trailing whitespace would be included because they'd be
evaluated as True.

```
Example results:
"" becomes bool("") -> False
" " becomes bool("") -> False
"A " becomes bool("A") -> True (but still returned "A ")
```

To fix this and preserve the behavior, this runs map first then filters
out anything that is now Falsey. So that last example returns: "A"