Fix rare flakiness from ordering differences (#37)
* Fix rare flakiness from ordering differences

We occasionally saw this test fail due to the hashes ending up being
different between the two concept objects. It tursn out the preferred_label
ordering is not deterministic, so will occasionally result in the hashes
being different. To flush out the issue I've added more items to the
preferred_label field, which caused the error to happen much more
frequently. I've then added a simple alphabetical sort to ensure these
are always the same order.

* Update tests/test_concept.py

Co-authored-by: Harrison Pim <hello@harrisonpim.com>

---------

Co-authored-by: Harrison Pim <hello@harrisonpim.com>