Fix pytorch for macos intel machines
Pytorch has deprecated support for intel macs.[^1][^2] Meaning it is not
possible to setup the environment on these machines with a pytorch
version over 2.2.2. Rather than forcing everyone else to stoop down to the
level of these poor souls, this aims to only fix the version on the
specifically affected machines.

[^1]https://dev-discuss.pytorch.org/t/pytorch-macos-x86-builds-deprecation-starting-january-2024/1690
[^2]https://github.com/pytorch/pytorch/issues/114602