Move config out of global scope
The config object was orignally static and therefore could be global.
We've added code to it that connects to prefect cloud, and having this
at the global level means it wants to run on import, which breaks our
tests.

Moving instead to a more function specific scope gives more flexibility
generally, but also makes it much easier to test and mock the config
object safely, especially the prefect block.