Only import `EmbeddingClassifier` when called
We previously made it possible to install dependencies without embeddings
libraries, but missed updating the module import behaviour. The previous
version would have imported the `EmbeddingClassifier` and therefore also
`sentence_transformers` from that file. Here we change the module
behaviour, so that we only actually import the `EmbeddingClassifier` when
we are explicitly using it.

This approach makes sure both of the below import statments still work.
The latter requiring the `embeddings` group to have been installed with
poetry:
```
from src.classifier import Classifier
from src.classifier import EmbeddingClassifier
```

This has the added bonus of saving a few seconds of start up time for
anything not using it (as sentence_transformers can take 5-30 seconds
to import).