Proof of concept, concept extraction
Initial script to pull concepts from the text of a document. This
implementation uses KeyBert. A Bert based LLM that works by comparing
substrings against a whole piece of text. In this case key phrases within
a text_block. From here this script then counts the most frequently
occuring and prints the results as a table.

There are lots of ways we could take this further, the most obvious ones
to me being:
- to run it over more documents
- to consider longer substrings than two word substrings
- explore other methods available from the same library, such as KeyLLM