Fix typing for infer arguments
As typer considers these a list and we convert to None later if the list
is empty. Also adding the same treatment for classifiers for consistency.
Currently this will cause the flow to fail where previously it caused it
to do a no-op run as the empty list basically just skipped all of the
flows functionality, so this seems more sensible