Adding the ability to provide partial updates for specific documents. (#101)
* Adding flows with tests to enable the partial updates of text blocks / passages in vespa with concepts.

* Adding some updates post pr comments - documentation.

* Raise ValueError on no passages for concept.

* Adding context manager for temp dir.

* Updating from Path to os.

* Updating code for updating a passages concepts with the new concept

* Refactoring due to line lengths.

* Replacing semaphore with prefect concurrency.

* Renaming functions.

* WIP.

* WIP.

* Fixing test_get_passage_for_concept.

* Fixing tests and resolving bugs post conversion to labelled passages.

* Minor refactors.

* Adding stronger asertion to test.

* Updating the ssm parameters to be decoded post retrieval.

* Updating the parameter names for the vespa ssm params post deploy.

* updating the naming for the ssm params.

* Update typo in flows/index.py

Co-authored-by: Jesse Claven <jesse.claven@me.com>

* Updating the ssm parameters to be decoded post retrieval.

* Updating the parameter names for the vespa ssm params post deploy.

* updating the naming for the ssm params.

* Clearer typing for inference classifier spec

Changes the classifier spec to use a pydantic model. Meaning we get some
nice documentation in the prefect user interface and less serialisation
challenges from the tuple based input. This should be useful for cli usage
(more to follow), and also for webapp usage

* Feature/pla 238 add extra fields to stored labels (#92)

* Adding parent concept to the labelled passages.

* Refactor.

* Updating the metadata to contain all information on the concept.

* Adding text block id to the metadata.

* Add the timestamp of inference to the metadata.

* Removing text block id from metdadata.

---------

Co-authored-by: Mark <mark@climatepolicyradar.org>

* Resolving merge conflict.

* Ensure api key is a secret str type

* Pass actual value of wandb api key to wand login

SecretStr inherits from str so probably would have been valid, however
wandb is quite strict about it. See:

https://github.com/wandb/wandb/blob/main/wandb/sdk/wandb_settings.py#L571-L572

As a bonus this may also solve some linting issues seen in some local
setup!

* Add manual Inference deployment trigger

This adds an easy interface to trigger inference in any environment.
With this any number of classifiers and documents can be specified. For
example:

```
just infer-sandbox -c Q787:v0 -c Q22 -d CCLW.executive.10002.4495 -d CCLW.executive.10126.4646
```

Or all documents:
just infer-sandbox -c Q787:v0 -c Q22
```

* Ensure empty documents are set to None

* Make env specification more consistent for infer

Using the matching enum used elsewhere. Also as part of this, changes
the prod deployment to use the consistent naming to how we're describing
it elsewhere

* Fix typing for infer arguments

As typer considers these a list and we convert to None later if the list
is empty. Also adding the same treatment for classifiers for consistency.
Currently this will cause the flow to fail where previously it caused it
to do a no-op run as the empty list basically just skipped all of the
flows functionality, so this seems more sensible

* Add infer to poetry script

* Bugfix for the ssm param test.

* Adding the correct concept test data.

* Adding the wandb mock as a fixture.

* Adding a generator for s3_paths.

* Adding a valuerror if the labellers array is not len==1

* Updating the comment around concepts being attached to labelled passages.

* updating the test data to be more representative.

* Adding test for the s3 generator.

* Typo.

* Raise a value error should both of the s3 parameters be set.

---------

Co-authored-by: Mark <mark@climatepolicyradar.org>
Co-authored-by: Jesse Claven <jesse.claven@me.com>
Co-authored-by: olaughter <fred@climatepolicyradar.org>