Add option to run on all available classifiers
For most automated runs we will want to use all available classifiers in
a given environment. For how we set this up in weights and bias, this
means finding all of the models with the wikibase_id name convention,
checking their metadata and selecting one that matches the environment
being run in.