refactor(index): Use equivalent interface with inference
We use the equivalent interface from inference for indexing. This has
the benefit of a. being familiar and b. reasoning about at the same
abstraction of document IDs and classifier specifications instead of
S3 keys/prefix, that are an incidental technical implementation.

Allows running on all classifiers and all documents or specific
classifier and document IDs.

Originally proposed on Linear [1].

[1] https://linear.app/climate-policy-radar/issue/PLA-287/create-interface-between-inference-and-index#comment-ad3306b1

TOWARDS PLA-287