Skip wandb org deduction
We were seeing a weird error in the deployment that could not be
recreated locally. The wandb code that was erroring was the code that
tried to deduce the organisation before calling ArtifactType. We don't
need wandb to do this for us, so can skip this step and go straight to
ArtifactType.