fix: wrap wandb list models in run to avoid timeouts
For whatever reason, the hidden 'magic' of wandb's internals was causing
the download model part of inference to hang on the wandb init. This only
seemed to be a problem in the following scenario:
- When running as a prefect deployment
- When running the flow concurrently
- Having used the wandb.Api interface previously in the flow

Interacting with the api as a wandb run seems to resolve the issue,
presumabley as it allows wandb to clean up any existing connections.