Add intermediary model identification stage
We where having problems with the wandb sdk trying to deduce all relevant
models as part of the inference prefect flow. The code worked outside of
prefect, but it would hang on wandb.init during classifier download if
the wandb api had already been used. Although only when running it as a
prefect deployment. Its not clear exactly what combination of prefect,
concurrency, wandb usage is causing this.

Resolving the issue was taking too long, so as a workaround this introduces
an intermediate step, of extracting the model identification functionality
into a script that can be ran manually. The output is stored in yaml and
can then be read by the prefect flow.

This has the obvious downside of needing to be updated periodically, but
I'm hopeful we'll be able to revisit at some point. On the plus side
having yaml files containing what we expect to run does also make it
extra clear, which will be nice in the first uses, so maybe this is a
better start point anyway.