fix: Use latest RulesBasedClassifier model for all promoted
For #188, I only did the deploy script for one classifier, Q956, that
used this architecture.

When the weekly pipelines ran on Monday, Q412 also failed [1].

I ran the deploy script for all concepts that a. have negative labels
in my current cache and b. are in the production classifier specs.

Namely, I ran these 2 commands in order:

1. `$ AWS_REGION=eu-west-1 poetry run deploy new --to-aws-env staging --wikibase-ids Q992 --wikibase-ids Q955 --wikibase-ids Q787 --wikibase-ids Q766 --wikibase-ids Q765 --wikibase-ids Q412 --no-get`
2. `$ AWS_REGION=eu-west-1 poetry run deploy existing --from-aws-env staging --to-aws-env production --no-get --no-train`

In hindsight, this went over _all_ the classifier specs. in production,
but alas.

[1]  https://app.prefect.cloud/account/4b1558a0-3c61-4849-8b18-3e97e0516d78/workspace/1753b4f0-6221-4f6a-9233-b146518b4545/runs/flow-run/b56dfb8d-87e2-47cb-b56f-db72ebe88c3c?entity_id=0ee696e9-66c4-4cb9-8736-e4707f56da1d

FIXES PLA-380