feat(indexing): Use concurrency across containers
We've faced the same problem as when doing inference. This changeset
means that we can either run the partial updates themselves as
deployments (default) or as functions.

This optionality was primarily done so that we can have increased test
coverage, without Prefect trying to actually run the deployments, for
the functions that do that. Still, I've needed to skip 1 test.

The function layout perhaps isn't amazing, but, there were many ways
to do this all, and this was one of them.

TOWARDS PLA-390