refactor(indexing): Batch conversion to concepts
Looking at the current running flow [1], it's up to 44085 log lines.
It's roughly 2 per labelled passage. That's a huge amount of function
calls to do linearly.

Originally I tried moving conversion batching to within the new
deployments we run. That didn't pan out as the design of this workflow
is _almost_ unmaintanable, and when I tried refactoring it all, it
seemed untenable at the time.

For now, the batching is done in this spot. We can try again at
improving it.

[1]
https://app.prefect.cloud/account/4b1558a0-3c61-4849-8b18-3e97e0516d78/workspace/1753b4f0-6221-4f6a-9233-b146518b4545/runs/flow-run/a71b5a1c-58d7-4549-8baf-2fc8ddb9eed2?entity_id=a71b5a1c-58d7-4549-8baf-2fc8ddb9eed2
[2]
```
converting labelled passage (ID: `383`) to Vespa concept
converting 1 spans to concepts
```

TOWARDS PLA-390