fix(indexing): Restrict parameters to less than Prefect's limit
I did consider several places to do this. It could've been before the
"original" batch size, but then we're back to the previous problem of
iterating over that big function.

This seems like the reasonable solution for now.

There's 2 caveats, 1. about the calculation for the likely number of
max concepts, and 2. about possibly expanding past the "original"
batch size. For 2., what if snipped out the excess, and inserted it as
the next batch? That seemed risky, for now.

FIXES PLA-397