Add initial mock test coverage for WikibaseSession
I was keen to have this in order to write tests for the flow that will
copy concepts from wikibase to S3. It's a relatively light touch mock
interface that could be improved, particularly around:

1. What the fixture data looks like, for this initial implementation I
just ran each query and used part of the response.
2. Whats actually being tested, I wanted to add something for `get_concept`
and `get_concepts` as this is what I'm going to build on, but for now
these are light touch. Other endpoints are even lighter but I'm hoping
this adds some foundation that could be built on.