feat(indexing): Calculate and update concepts' counts on family documents
Since the generators are deterministic, and go document × passages, we
can accumulate the concepts during the batches, document by document,
and then do the update.

The important part is where we track which document we're on, and the
accumulation of concepts' counts. We then save them to S3.

There's a few hairy parts of this, but I think it's sufficient for now.

CLOSES PLA-414