feat: Update family documents' concepts counts (#239)
* feat: Update family documents' concepts counts

Group all the outputs, that are per document × concept, so we have a
list of all concepts by document.

Then, collect all of them together, and do 1 partial update per
document.

There's a lot of fixtures. I copied a ton from our production S3
bucket. I renamed a few so there's some document import IDs that are
present in all of the concepts.

TOWARDS PLA-415

* Load all concept fixtures into mock s3

This automatically picks up concept count fixtures and adds them to the
mocked bucket, letting us use them for tests

* Use vespa response method for checking status code

Taking advantage of this method provided by vespa for a very minor
simplification

* Refactor handling of concept count object

Turns out that vespas library doesnt need this serialised in advance but
will do that for us when passing a dictionary. Moving the transformation
up out of the update function also means we can return it to the
gathered tasks rather than the count object, this is nice because its the
exact object we sent to vespa so we can build on this to check its worked

* Add initial test for loading concept counts

Tests the main subflow of the concept_counts indexing

* Parameterise tests

* Ensure tests cover multiple concepts counts

---------

Co-authored-by: olaughter <fred@climatepolicyradar.org>