Add default property relationship values
These are defined in the env.example but where missed for the env variables
when setting up the pipeline copying concepts from wikibase to s3, the
result was that these objects ended up without the relationships.

Adding them to the pipeline was the most obvious option, but I figured
having them as defaults on the interface client might save us from doing
this again, while also allowing for flexibility.