feat: De-index results from S3 and Vespa
The inverse of indexing, and I guess, inference too.

It removes the results from S3 and Vespa.

It avoids a scan of Vespa by first collecting what should be
deleted in Vespa, by analysing what should be deleted in S3.

It uses the same interface as the other pipelines.

This would be triggered from the Wikibase to S3 backup automatically
and by DSci after updating the classifier spec manually.

FIXES PLA-469