feat: Add de-indexing pipeline
This pipeline reads what's the state in S3, and then undoes that from
Vespa (hence, the "de-index") and then removes the S3 objects as well.

It uses the same interface as other pipelines, except that you must
specify the classifiers specs, as we don't want to ever allow a full
removal of all concepts.

This approach avoids a full scan of Vespa, to remove a concept.

**Resiliency on partial updates failures**

De-indexing accounts for if updating a document passage fails. When removing a concept from the document passage fails, it'll still be counted, and written to S3.

The usage here is intended to re-run the de-indexing pipeline again, for the concept(s) and family document(s), if the failure wasn't related to the pipeline itself. This could be a transient network failure. This very generously could be called idempotency.

**I don't see changes for removing concepts from concepts counts**

The concepts counts-related pipeline does a wholesale overwrite, after it's collected the concepts counts from S3. That means, a concepts counts object for concepts will be deleted in de-indexing, and thus won't be seen during the counting pipeline, and will be removed from the family document.

**There's so much duplicated code now!**

Yes, there is. Once it all is working, we can look at refactoring it to be DRY.

TOWARDS PLA-481