feat: Clean-up and de-index concepts (#302)
* feat: Add de-indexing pipeline

This pipeline reads what's the state in S3, and then undoes that from
Vespa (hence, the "de-index") and then removes the S3 objects as well.

It uses the same interface as other pipelines, except that you must
specify the classifiers specs, as we don't want to ever allow a full
removal of all concepts.

This approach avoids a full scan of Vespa, to remove a concept.

**Resiliency on partial updates failures**

De-indexing accounts for if updating a document passage fails. When removing a concept from the document passage fails, it'll still be counted, and written to S3.

The usage here is intended to re-run the de-indexing pipeline again, for the concept(s) and family document(s), if the failure wasn't related to the pipeline itself. This could be a transient network failure. This very generously could be called idempotency.

**I don't see changes for removing concepts from concepts counts**

The concepts counts-related pipeline does a wholesale overwrite, after it's collected the concepts counts from S3. That means, a concepts counts object for concepts will be deleted in de-indexing, and thus won't be seen during the counting pipeline, and will be removed from the family document.

**There's so much duplicated code now!**

Yes, there is. Once it all is working, we can look at refactoring it to be DRY.

TOWARDS PLA-481

* feat: Clean-up and de-index concepts

We now wipe the concepts from S3 (the CDN).

Along with this, we trigger the new de-indexing pipeline, to remove
the concepts' inference and concepts counts objects, and labelled
passages (spans) in Vespa.

TOWARDS PLA-481
TOWARDS PLA-469

* Resolving failing tests.

* Reverting some minor changes from the merge.

* refactor: Share code between flows for systems boundaries (#303)

* refactor: Share code between flows for systems boundaries

I found it easy to think about this as the boundaries, and where we
traverse them. This isn't the final step in sharing, but is a good
start imo.

TOWARDS PLA-481

* Refactor/further deduplications (#311)

* Attempting further deduplications.

* Reverting change to the partial updates.

* Moving functions for PR clarity.

* Further refactors.

* Deduplicating further by parameterising on partial update callable.

* Fixing failing tests.

---------

Co-authored-by: Mark <mark@climatepolicyradar.org>

---------

Co-authored-by: Mark <mark@climatepolicyradar.org>
Co-authored-by: NextGenEng <58440325+THOR300@users.noreply.github.com>

---------

Co-authored-by: Mark <mark@climatepolicyradar.org>
Co-authored-by: NextGenEng <58440325+THOR300@users.noreply.github.com>