fix: Deserialise per item
Nothing notable to say, really. This is the way the data is serialised
at the present. I got a ton of failures in Prefect[^1].

Ideally, we're not saving them in 2 different ways, especially one
that's broken. This PR at least gives as a nice migration when fixing
the other way.

TOWARDS PLA-481

[^1]: https://app.prefect.cloud/account/4b1558a0-3c61-4849-8b18-3e97e0516d78/workspace/1753b4f0-6221-4f6a-9233-b146518b4545/runs/flow-run/1210c142-de95-4ae2-9201-b531244a8b69?entity_id=5cae68f4-8db7-4142-a34a-4a67ea7bd63b