fix: Properly timeout on Prefect
We fundamentally misunderstood what the timeout parameter was for.

> By default, this function blocks until the flow run finishes
executing. Specify a timeout (in seconds) to wait for the flow run to
execute before returning flow run metadata. To return immediately,
without waiting for the flow run to execute, set timeout=0.[^1]

It’s a timeout for returning metadata, not for execution to finish.
E.g. by having 3600, we’re saying “no matter if the `FlowRun` is still
executing, return the metadata, and let the `FlowRun` continue on
until it finishes, with no actual timeout”.

For an incident[^2], we were confused as to why they hadn't timed out!
We actually had been confused before too, and we did set a timeout, or
increased it, I don't remember which. We didn't realise this then due
to our general sporadic work across many projects.

Now, we properly wait![^3]

An alternative approach is to annotate the decorator:

```python
from prefect import flow
import time

@flow(timeout_seconds=1, log_prints=True)
def show_timeouts():
    print("I will execute")
    time.sleep(5)
    print("I will not execute")
```

The change I'm proposing felt more suitable with the current code.

What if a task actually goes over? I'm not sure. I already know that
this timeout will be too short for really big documents.[^4]

[^1]: https://docs-2.prefect.io/latest/api-ref/prefect/deployments/deployments/#prefect.deployments.deployments.run_deployment
[^2]: https://climate-policy-radar.slack.com/archives/C08LN4VA2D8/p1743668576959819
[^3]: https://docs-2.prefect.io/latest/api-ref/prefect/flow_runs/#prefect.flow_runs.wait_for_flow_run
[^4]: If you look at
https://app.prefect.cloud/account/4b1558a0-3c61-4849-8b18-3e97e0516d78/workspace/1753b4f0-6221-4f6a-9233-b146518b4545/runs/flow-run/3a6dbe77-bebb-4788-89d2-f581ce2e948e?entity_id=5bedd38b-8432-4f1b-b792-3a71bbcb4eb7,
you'll see most are relatively quick, but a few are >3, or >6 hours.