feat: Add Result type
These adds a classical Result type. It carries the _usual_ benefits.

Here's a few of them:

- Exceptions should be exceptional, not known
- Allows you to continue on (via `map_`) without excessive error
  handling at that point in time
- Makes it more clear what a function can return, as opposed to
  invisibly, possibly, returning an exception
- Makes it explicit what the outcome states, and let's us reason about
  them
- You can optionally have anonymous or typed errors
- Type checking on possible returns from functions becomes checkable

This is intended as an experiment, though as you can see from the
example here, it may not be a tedious one to rollback, unless someone is
very good at mass refactorings.

As a bonus, if I've introduced the unit type, that's analogous to a
`None` value in Python. It's in Haskell, Rust, etc. This is purely an
extra, and not required at all. It's hard for me to explain, but after
having first come across it in Elm, I found the "no information" that it
represents, much more clear and semantic than a `None` type. It also has
the benefit of being explicit, since `None` can be implicit, like
functions with no returns.