fix: Use non-memory leaking aiohttp
When building the image, I noticed this warning:

```
12.48 Warning: The file chosen for install of multidict 6.3.2 (multidict-6.3.2-cp310-cp310-manylinux_2_17_aarch64.manylinux2014_aarch64.whl) is yanked. Reason for being yanked: Memory leak: https://github.com/aio-libs/multidict/issues/1131#issuecomment-2787514071
```

Poetry also warned about it:

```
Warning: The locked version 6.3.2 for multidict is a yanked version. Reason for being yanked: Memory leak: https://github.com/aio-libs/multidict/issues/1131#issuecomment-2787514071
```

I've explicitly added aiohttp as a dependency so we can keep it _more_
up-to-date. It now is a version that uses a fixed[1][2] version of multidict.

[^1]: https://github.com/aio-libs/multidict/issues/1131#issuecomment-2787514071
[^2]: https://github.com/aio-libs/aiohttp/blob/1acf8c94e7b5ad7d1a2ce6cfec0937abbe43a6eb/requirements/multidict.txt#L5