fix: Use the primary model version, not the latest
Until now, we were effectively relying on the order of model artifacts,
to determine what the _right_ verion is to use. The implicit ordering
reliance is for newest to oldest.

As part of promotion, the AWS env. is attached to a model version, to
indicate this _the_ one to use by default.

This change now looks for the model version that has the appropriate AWS
env. label, that indicates primariness.

FIXES PLA-584