Make it easier to audit gaps in inference results (#393)
* Simplify interface

The protocol isnt needed by the boto paginator, and the prefix is always
the same

* Use map to simplify environment specification

Rather than having to pass in the whole path to the yeml, this makes it
just need to be the environment

* Refactor to isolate single spec check functionality

This is intended to lay the groundwork for running each spec in parallel.

* Add concurrency

Runs across multiple classifier specs at the same time.

* Simplify logging

Aims to reduce some repetition in the logging as most of the s3 path will
always be the same.

* Keep file names in lookup

This is intended to let us write these out in the future.

* Add write out option

* Refactor write_result into own function

* Add test coverage

* Add env to write out dir path

* Fix aws env type