fix: Safely access shared data across Vespa feed callbacks
Whilst we don't have reproducible steps for the memory issue, and thus
I'm not certain if this will fix the issue, it has at least led me to
identify this issue.

While list operations are atomic[^1], I have some doubts about relying
on the bytecode ordering, and they do suggest to use a mutex if in
doubt, so here's a mutex in the form of a lock.

TOWARS PLA-573

[^1]: https://docs.python.org/3/faq/library.html#what-kinds-of-global-value-mutation-are-thread-safe