Log print statements
Prefect was missing log statements that where found in ecs. This is due
to the function `s3_paths_or_s3_prefixes` having print statements rather
then use logging. Elsewhere we have adopted print statements for prefect
flows because they are cleaner and seem to behave better with concurrency
so I have switched the main flow here to do the same. This will then
allow the s3 functions print statements to be logged