Add precision to grouped query
Because of the way vespa works with querying we where losing passages in
the response. This is because by default the way vespa handles searches
across nodes and pools the results is optimised for performance not
accuracy.

In our case here, we kinda really need accuracy and this can be changed
by overriding the default "precision" parameter in the grouping query.

See:
- https://docs.vespa.ai/en/reference/grouping-syntax.html#precision
- https://docs.vespa.ai/en/grouping.html#performance-and-correctness