refactor: Allow flexible async coroutines
A batch of tasks means that the batch has to wait until all tasks
finish, before starting the next batch. That means the coroutines are
_inflexible_.

Instead, use a queue AKA lock AKA semaphore, so that the concurrency
is managed by there being _space_ in the semaphore. E.g. a semaphore
of 8 means that there's 8 concurrent tasks. A task blocks until it
acquires a lock, from the semaphore.

TOWARDS PLA-615