Add initial inference results aggregation flow  (#425)
* Add initial inference results aggregation flow

This takes the raw inference output, aka "laballed_passages", it groups
them by passage and writes the results as `VespaConcept` objects back to
s3 under "inference_results".

There's more that can be done for speed on this flow, but I'd like to
take it for some spins without concurrency before adding any.

We will also need to add a sensible default for not specifying document
ids at some point, but again felt we could air it out a bit before figuring
that out.

* Add deployment for aggregation flow

* Make output path configurable

Adding this to the config means we can also change it when necessary

* Update typing with text block alias