refactor: Use stricter type hinting for `str`s and constants
This came out of me, in a code review, asking for more of these
`TypeAlias`s to be added, and wondering if they're effective for
static checks, and not just making function signatures more human readable.

Turns out, they weren't good for the former[^1]. There is a little
boilerplate with having to instantiate the type, but alas, I feel the
potential type checker warnings is worth it. We'll see though, and
maybe there's a way to avoid this.

[^1]: https://climate-policy-radar.slack.com/archives/C078S46DJFP/p1748337885539729