Add fallback to aggregate using all document ids
This adds the behaviour to run on all available inference outputs when
None are specified. To do this it gathers all the document ids in the
raw inference results. This way we have the option to not specify what
documents we are aggregating on, which is useful when we want to run on
everything.