refactor: Index both document passages and family documents together
 I've counted all concepts counts for all the updates that worked, and
 the failures are reported.

One thing I've experienced here is that I feel by not doing things the
Prefect way[^1], that I'm fighting Prefgect. For example, we aim to be
resilient and do _as much as possible_, amongst the failures.

This shows up in how we do tasks gathering and then raise _dumb_
exceptions, to get Prefect to show an overall, failed state. Instead,
if we were more Prefect-first, I think we could avoid this munging of
success and failure returns and reduce code as well.

As from PLA-622 and our discussion on naming (indexing both things),
I've used that naming around here.

TOWARDS PLA-622

[^1]: https://climate-policy-radar.slack.com/archives/C078S46DJFP/p1748959406451039