feat: duplicate latest run documents for backfills
The idea here is to have all the latest ran aggregations compiled in a
single place. Previously, all the documents to be drawn on for indexing
would need to be within a single run identifier. Most of the time this
is really useful and lets us only run on what has just been completed.
But there are also times when we need to rerun on everything, or when we
want to review what the current state should be. In this case it helps to
also maintain all of the last aggregated documents.

We previously had some bad experiences with `latest` with models, but I
think this is different enough from that situation that its worth having.