refactor: Remove unused indexer
This has been replaced by the indexer from aggregated results.

The final concepts counts are the same, as the S3 prefixes/paths
parameters refer to labelled passages that aren't in the base fixture
for Vespa. That means it's testing that de-indexing doesn't remove
other things, but it's not testing that de-indexing is removing
certain things. Tests for that can be added later.