refactor: Use task mapping and log Vespa updates
Task mapping has worked well recently in inference aggregation. It's
used here as one way of getting better observability, as it's unclear
right now which `index_all` runs fail.

Along with this, it's unclear which Vespa updates go wrong, so this
improves that, and also there's additional logging to understand the updates.

TOWARDS PLA-658