fix: Reduce task concurrency
I hit Prefect's rate-limit[^1]. The concurrency is _batches of
documents × number of tasks per batch × number of connections in Vespa
connection pool_.

Previously in #502, where the connection pool was instantiated was
changed, and it was predicted that these values would need to be tweaked.

There's one parameter I've removed too, that wasn't really needed,
since the concurrency of document passages being updated was locked to
the max of connections in the Vespa pool, so it wasn't helpful to have
updates concurrency > max connections.

NB: Flows and task have `with_options`. Unfortunately we can't use it
on flows that are being run as deployments.

TOWARDS PLA-658

[^1]: https://app.prefect.cloud/account/4b1558a0-3c61-4849-8b18-3e97e0516d78/workspace/1753b4f0-6221-4f6a-9233-b146518b4545/runs/flow-run/068598aa-f6cd-71f7-8000-bea53d5dd26c?entity_id=068598b6-0264-7944-8000-dd211b2a46a2&state=failed&entity=flowRuns&entity=taskRuns&entity=events&entity=logs&entity=artifacts