feat: spec reader interface for specs with more data
We are moving to having specs with more data included in them. This starts
an interface for working with these new file structures. There was plenty
of code existing spread out that interfaces with the old files, this
begins to consolidate that for the new files.

More work will be needed to update before and after this point in the
lifecycle (i.e. training, inference, etc). It seemed like a good place to
start as these tie everything else together.