refac: remove `only_run_on` for now
This was exploratory while figuring out what we might want from these
specs in the future. As we will want to filter certain document groups.
I think this proves the potential of using them this way, but is also
beyond the scope of the current change so have removed it for now.