feat: Add progress reporting
Use the Prefect progress kind of artifact for places where it either
takes generally a while and/or there's a lot of things to enumerate through.

I've tried to make it as _for free_ as possible, so it's in the
sub-flow mapping function, and I've created a gather equivalent.

FIXES PLA-610