refac: flatten track upload conditionals
Having moved out the finish we can combine these together, this is safe
because track is required for upload anyway and we validate this at the
start of the script. This also flattens the version bumping so that the
version is looked up regardless. This makes the local models more aligned
with uploaded models, and allows some simplification by removing a
conditional.