test: always skip the tests that contact argilla
These tests were actually creating a dataset in argilla. Judging by the
docstring it is not intended to be run on a normally in tests and that
shows in ci where we ignore it. However when running locally its easy not
to ignore this module, I've therefore set it to always skip for now.
I've also ensured local env variables for argilla are not picked up on
by defining them for pytest with test values.