feat: include env in the wandb `artifact` alias
The model connection between artifacts and s3 stores is a one to one,
which means unless we revisit this our models will all have a coupling
with a specific environment. This is now the only tangible difference
between versions of a model. This adds the env as an alias to each model.

The benefits to this is are:

* Clarity for humans when reviewing wandb which model is for which
environment
* Easily identify the environment for a model programmatically in a more
dynamic way.