refactor: Remove `Across` code.
We ended up leaving across account promotion unimplemented. Now that we
are changing promote it makes sense to remove it so that we don't have
to maintain something we are not using. We probably will want better
handling of accounts and train/promoting one day. At which point if we
decide we still want to continue with the approach we had started here
we can still reach into the commit history.