ci: add stale pr detection (#563)
We have a number of stale PRs across our codebase, I'd like to find a
way of keeping these cleaned up and am interested in trying this out here
in the knowledge graph repo first. I've taken 40 days as the amount of
time before a PR should actually be closed for now. My view is that after
this time the PR is probably never going to be merged because the rest
of the repo has changed, so it either should have gotten merged earlier
or needs to be reopened. The specific timing could be adjusted though.

Note that for exceptions the label `stale-exempt` can be added to a PR.

See for reference: https://github.com/actions/stale