feat: Support multiple model versions for concepts' classifiers
This is for the RFC[^1] around hot-swapping concepts. The first
solutions required deploying changes to the schemas and/or SDK.

This bypasses that, and enables flexibility in querying and in
updating Vespa's content.

**Enabled Functionality**

I'll go through some funcionality that this enables.

_Configure versions_

- Find all documents that have certain concepts
- Find all passages that have certain concepts
- Find all documents that have certain concepts with a specific version
- Find all passages that have certain concepts with a specific version

With this, the Policy team could select some pre-release versions to see.

_Spans as first-class_

They were previously in the concept. They're now slightly more
first-class. This data model makes it so the spans are specific to the
output from a model. Not all versions of the model will return the
same spans.

I think this would make it easier for the Application team to reason
about multiple, and possibly overlapping, spans. I was trying the
webapp during this, and found it tricky to understand spans visually.

_Concepts as first-class_

With them as first-class schemas, we could now build pages around them.

We could also query by their names. Say someone is interested in
fishing. They could query for `fish` and get ≥ 0 concepts. We could
then find all the documents and passages that match it.

It could also be possible to do this in one step by adding the concept
name's to `concepts_instances`.

**Data Model**

NB: `concepts_instances` could be renamed `concepts` to replace the
existing, or `concepts_v2`.

A document or passage now has concepts as a map instead of a list.
Previously, a list of concepts could include the same concept twice,
to account for multiple spans. This flips that on its head so it's
concept→versions (aka IDs).

The similar approach of _flat_ IDs is used for parents and models' IDs.

The idea would be to get a passage, which includes all the spans by
model version. You then access them by the latest model ID, or, by a
specific model ID, such as a pre-release one.

NB: The latest ID version doesn't even need to be in the documents and
passages too. It could a global thing, à la search weights. That means
we'd only need to update it in 1 place. I'm not sure of the querying
or search profiles complexity, but perhaps no more than the search weights?

NB: Spans could become a separate schema. I'm not entirely aware of
grand aspirations of spans.

**What about our pipelines?**

I'm fairly certain this could all be done at the index pipeline.

**Surely there's downsides?**

See: **Benchmark**.

There's some data duplication, around satisfying the lack of joins,
but no more than we already have.

**Test**

I have a simple script that goes through some use cases. I recommend
running these 2 shell commands. After that, try out some queries
yourself via the Vespa CLI.

```bash
$ just vespa_dev_down vespa_dev_setup vespa_feed_data
$ ./test_vespa_queries.sh
```

**Bencmhark**

I'm not 100% sure about the performance, as I don't have experience
with Vespa as I do with a SQL DB, and I can point out a lot more bad
data models and bad queries.

[^1]: https://www.notion.so/climatepolicyradar/RFC-Hot-swapping-concepts-1de9109609a48053b017d5745c8fef67

FIXES PLA-774