test: scope test prefect server to test flows
When running `pytest tests/test_promote.py` with this line tests take
17.6 seconds. Removing it and they take 1.2 seconds. All of our prefect
test flows live in `flows/`, which has its own conftest file that spins
up this prefect server. Meaning any time we are running just on a module
outside this repo we are losing ~15 seconds per test run for no reason.

I don't expect this to have any benefit in ci, but when running non-flow
tests locally this should save time. Note that if we ever have flows
outside of the flow directory we'll now need to ensure they have the
harness.