fix: check for project artifact in registry collection
When looking up the artifact we where using the version, the projects &
registry versions are not kept in sync with each other meaning there is
an edge case that would cause this to fail. Now that we are looking to
use the classifier id instead of just the version it can be made clearer
what we're looking for.