feat: add script to mark classifier metadata
This adds a classifier metadata inference that allows items to be written
and removed from a `dont_run_on` key. Downstream this will be used to
filter ids.

The script is designed to be extendable, currently focusing on filtering
out sources, but we could further consider adding compute settings here,
etc